Rename policywizardv1 to policysetv1 and improve URL handling
