+++ cyral/internal/policyset/v1/constants.go
@@ -1,7 +1,7 @@
-package policywizardv1
+package policysetv1
 
 const (
-	resourceName     = "cyral_policy_wizard_v1"
+	resourceName     = "cyral_policy_set"
 	dataSourceName   = resourceName
 	apiPathPolicySet = "v1/policyWizards/policySets"
 )
