+++ cyral/internal/policyset/v1/datasource.go
@@ -1,8 +1,8 @@
-package policywizardv1
+package policysetv1
 
 import (
-	"context"
 	"fmt"
+	"net/url"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
@@ -11,71 +11,62 @@ import (
 	"github.com/cyralinc/terraform-provider-cyral/cyral/core/types/resourcetype"
 )
 
-var resourceContextHandler = core.DefaultContextHandler{
-	ResourceName:                 resourceName,
-	ResourceType:                 resourcetype.Resource,
-	SchemaReaderFactory:          func() core.SchemaReader { return &PolicySet{} },
+var dsContextHandler = core.DefaultContextHandler{
+	ResourceName:                 dataSourceName,
+	ResourceType:                 resourcetype.DataSource,
 	SchemaWriterFactoryGetMethod: func(_ *schema.ResourceData) core.SchemaWriter { return &PolicySet{} },
-	BaseURLFactory: func(d *schema.ResourceData, c *client.Client) string {
-		return fmt.Sprintf("https://%s/%s", c.ControlPlane, apiPathPolicySet)
-	},
 	ReadUpdateDeleteURLFactory: func(d *schema.ResourceData, c *client.Client) string {
-		return fmt.Sprintf("https://%s/%s/%s",
-			c.ControlPlane,
-			apiPathPolicySet,
-			d.Id(),
-		)
+		baseURL := &url.URL{
+			Scheme: "https",
+			Host:   c.ControlPlane,
+			Path:   fmt.Sprintf("%s/%s", apiPathPolicySet, d.Get("id").(string)),
+		}
+		return baseURL.String()
 	},
 }
 
-func resourceSchema() *schema.Resource {
+func dataSourceSchema() *schema.Resource {
 	return &schema.Resource{
-		Description:   "This resource allows management of policy sets in the Cyral platform.",
-		CreateContext: resourceContextHandler.CreateContext(),
-		ReadContext:   resourceContextHandler.ReadContext(),
-		UpdateContext: resourceContextHandler.UpdateContext(),
-		DeleteContext: resourceContextHandler.DeleteContext(),
-		Importer: &schema.ResourceImporter{
-			StateContext: importPolicySetStateContext,
-		},
+		Description: "This data source provides information about a policy set.",
+		ReadContext: dsContextHandler.ReadContext(),
 		Schema: map[string]*schema.Schema{
 			"id": {
 				Description: "Identifier for the policy set.",
 				Type:        schema.TypeString,
-				Computed:    true,
+				Required:    true,
 			},
 			"wizard_id": {
 				Description: "The ID of the policy wizard used to create this policy set.",
 				Type:        schema.TypeString,
-				Required:    true,
+				Computed:    true,
 			},
 			"name": {
 				Description: "Name of the policy set.",
 				Type:        schema.TypeString,
-				Required:    true,
+				Computed:    true,
 			},
 			"description": {
 				Description: "Description of the policy set.",
 				Type:        schema.TypeString,
-				Optional:    true,
+				Computed:    true,
 			},
 			"tags": {
 				Description: "Tags associated with the policy set.",
 				Type:        schema.TypeList,
-				Optional:    true,
+				Computed:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
 			},
 			"scope": {
 				Description: "Scope of the policy set.",
 				Type:        schema.TypeList,
-				Optional:    true,
+				Computed:    true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"repo_ids": {
 							Description: "List of repository IDs that are in scope. Empty list means all repositories are in scope.",
 							Type:        schema.TypeList,
 							Elem:        &schema.Schema{Type: schema.TypeString},
-							Optional:    true,
+							Computed:    true,
 						},
 					},
 				},
@@ -83,12 +74,12 @@ func resourceSchema() *schema.Resource {
 			"wizard_parameters": {
 				Description: "Parameters passed to the wizard while creating the policy set.",
 				Type:        schema.TypeString,
-				Required:    true,
+				Computed:    true,
 			},
 			"enabled": {
 				Description: "Indicates if the policy set is enabled.",
 				Type:        schema.TypeBool,
-				Optional:    true,
+				Computed:    true,
 			},
 			"policies": {
 				Description: "List of policies that comprise the policy set.",
@@ -123,8 +114,4 @@ func resourceSchema() *schema.Resource {
 			},
 		},
 	}
-}
-
-func importPolicySetStateContext(_ context.Context, d *schema.ResourceData, _ interface{}) ([]*schema.ResourceData, error) {
-	return []*schema.ResourceData{d}, nil
 }
