+++ cyral/internal/policyset/v1/resource.go
@@ -1,7 +1,9 @@
-package policywizardv1
+package policysetv1
 
 import (
+	"context"
 	"fmt"
+	"net/url"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
@@ -10,57 +12,77 @@ import (
 	"github.com/cyralinc/terraform-provider-cyral/cyral/core/types/resourcetype"
 )
 
-var dsContextHandler = core.DefaultContextHandler{
-	ResourceName:                 dataSourceName,
-	ResourceType:                 resourcetype.DataSource,
+var resourceContextHandler = core.DefaultContextHandler{
+	ResourceName:                 resourceName,
+	ResourceType:                 resourcetype.Resource,
+	SchemaReaderFactory:          func() core.SchemaReader { return &PolicySet{} },
 	SchemaWriterFactoryGetMethod: func(_ *schema.ResourceData) core.SchemaWriter { return &PolicySet{} },
+	BaseURLFactory: func(d *schema.ResourceData, c *client.Client) string {
+		baseURL := &url.URL{
+			Scheme: "https",
+			Host:   c.ControlPlane,
+			Path:   apiPathPolicySet,
+		}
+		return baseURL.String()
+	},
 	ReadUpdateDeleteURLFactory: func(d *schema.ResourceData, c *client.Client) string {
-		return fmt.Sprintf("https://%s/%s/%s", c.ControlPlane, apiPathPolicySet, d.Get("id").(string))
+		baseURL := &url.URL{
+			Scheme: "https",
+			Host:   c.ControlPlane,
+			Path:   fmt.Sprintf("%s/%s", apiPathPolicySet, d.Id()),
+		}
+		return baseURL.String()
 	},
 }
 
-func dataSourceSchema() *schema.Resource {
+func resourceSchema() *schema.Resource {
 	return &schema.Resource{
-		Description: "This data source provides information about a policy set.",
-		ReadContext: dsContextHandler.ReadContext(),
+		Description:   "This resource allows management of policy sets in the Cyral platform.",
+		CreateContext: resourceContextHandler.CreateContext(),
+		ReadContext:   resourceContextHandler.ReadContext(),
+		UpdateContext: resourceContextHandler.UpdateContext(),
+		DeleteContext: resourceContextHandler.DeleteContext(),
+		Importer: &schema.ResourceImporter{
+			StateContext: importPolicySetStateContext,
+		},
 		Schema: map[string]*schema.Schema{
 			"id": {
 				Description: "Identifier for the policy set.",
 				Type:        schema.TypeString,
-				Required:    true,
+				Computed:    true,
 			},
 			"wizard_id": {
 				Description: "The ID of the policy wizard used to create this policy set.",
 				Type:        schema.TypeString,
-				Computed:    true,
+				Required:    true,
 			},
 			"name": {
 				Description: "Name of the policy set.",
 				Type:        schema.TypeString,
-				Computed:    true,
+				Required:    true,
 			},
 			"description": {
 				Description: "Description of the policy set.",
 				Type:        schema.TypeString,
-				Computed:    true,
+				Optional:    true,
 			},
 			"tags": {
 				Description: "Tags associated with the policy set.",
 				Type:        schema.TypeList,
-				Computed:    true,
+				Optional:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
 			},
 			"scope": {
 				Description: "Scope of the policy set.",
 				Type:        schema.TypeList,
-				Computed:    true,
+				Optional:    true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"repo_ids": {
 							Description: "List of repository IDs that are in scope. Empty list means all repositories are in scope.",
 							Type:        schema.TypeList,
 							Elem:        &schema.Schema{Type: schema.TypeString},
-							Computed:    true,
+							Optional:    true,
 						},
 					},
 				},
@@ -68,12 +90,12 @@ func dataSourceSchema() *schema.Resource {
 			"wizard_parameters": {
 				Description: "Parameters passed to the wizard while creating the policy set.",
 				Type:        schema.TypeString,
-				Computed:    true,
+				Required:    true,
 			},
 			"enabled": {
 				Description: "Indicates if the policy set is enabled.",
 				Type:        schema.TypeBool,
-				Computed:    true,
+				Optional:    true,
 			},
 			"policies": {
 				Description: "List of policies that comprise the policy set.",
@@ -108,4 +130,8 @@ func dataSourceSchema() *schema.Resource {
 			},
 		},
 	}
+}
+
+func importPolicySetStateContext(_ context.Context, d *schema.ResourceData, _ interface{}) ([]*schema.ResourceData, error) {
+	return []*schema.ResourceData{d}, nil
 }
