+++ cyral/provider/schema_loader.go
@@ -16,7 +16,7 @@ import (
 	integration_teams "github.com/cyralinc/terraform-provider-cyral/cyral/internal/integration/teams"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/permission"
 	policyv2 "github.com/cyralinc/terraform-provider-cyral/cyral/internal/policy/v2"
-	policywizardv1 "github.com/cyralinc/terraform-provider-cyral/cyral/internal/policywizard/v1"
+	policysetv1 "github.com/cyralinc/terraform-provider-cyral/cyral/internal/policyset/v1"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/regopolicy"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository"
 	repository_accessgateway "github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/accessgateway"
@@ -57,7 +57,7 @@ func packagesSchemas() []core.PackageSchema {
 		integration_teams.PackageSchema(),
 		permission.PackageSchema(),
 		policyv2.PackageSchema(),
-		policywizardv1.PackageSchema(),
+		policysetv1.PackageSchema(),
 		regopolicy.PackageSchema(),
 		repository.PackageSchema(),
 		repository_accessgateway.PackageSchema(),
