+++ docs/data-sources/policy_set.md
@@ -1,12 +1,12 @@
 ---
 # generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "cyral_policy_wizard_v1 Data Source - terraform-provider-cyral"
+page_title: "cyral_policy_set Data Source - terraform-provider-cyral"
 subcategory: ""
 description: |-
     This data source provides information about a policy set.
 ---
 
-# cyral_policy_wizard_v1 (Data Source)
+# cyral_policy_set (Data Source)
 
 This data source provides information about a policy set.
