+++ examples/resources/cyral_policy_set/resource.tf
@@ -13,7 +13,7 @@ resource "cyral_repository" "myrepo" {
     }
 }
 
-resource "cyral_policy_wizard_v1" "repo_lockdown_example" {
+resource "cyral_policy_set" "repo_lockdown_example" {
   wizard_id    = "repo-lockdown"
   name        = "default block with failopen"
   description = "This default policy will block by default all queries for myrepo except the ones not parsed by Cyral"
