+++ templates/resources/policy_set.md.tmpl
@@ -6,6 +6,6 @@
 
 ## Example Usage
 
-{{ tffile "examples/resources/cyral_policy_wizard_v1/resource.tf" }}
+{{ tffile "examples/resources/cyral_policy_set/resource.tf" }}
 
 {{ .SchemaMarkdown | trimspace }}
